UI자잘자잘 수정

+++ src/main/java/com/kh/lahol/coffeeclass/model/vo/CoffeeClass.java
@@ -46,7 +46,7 @@ public class CoffeeClass{
 	private String clPhoto;
 	private String clWriter;
 	private String category;
-	private int avggrade;
+	private double avggrade;
 	
 	// report정보
 	private String rpReason;
